Add support for using Gitlab labels for entry type
For our workflow we use labels with a common prefix wit a separator
(e.g. 'CL:') for specifying the type of CL entry.

This commit adds support for using the label text after the separator
for the changelog entry type.

For this, two arguments are added for specifying the prefix and
separator (with defaults `CL` and `:`).

This information is passed on to the output logic.

To ease this, the `args` object is now passed directly to the action
functions.
This avoids defining redundant function arguments.