Add command `collect` for assembling MRs from issues
The newly added command `collect` walks through the list of issues,
associated with specified group milestones, collects all MRs linked
to the issues and assignes the specified project milestone to the MRs.

The idea behing this workflow is the following:
* We use group milestones for representing e.g. *Sprints*
* We use project milestones for representing release versions
* Issues are assigned to sprints
* Every issue has one or more MRs
* Only MRs shall be assigned to release milestones, issues stay
  untouched

The collect command produces this result.

The arguments needed are:
* Project ID (for getting / creating release milestone)
* Release milestone (i.e. project milestone)
* Group ID (for getting sprint milestones)
* Sprint milestones (i.e. one or more group milestones)